Use Long values for timestamps.

+++ src/nativeMain/kotlin/io/github/reblast/kpresence/rpc/Activity.kt
@@ -66,8 +66,8 @@ enum class ActivityType {
  */
 @Serializable
 data class ActivityTimestamps(
-  val start: Int? = null,
-  val end: Int? = null
+  val start: Long? = null,
+  val end: Long? = null
 )
 
 /**
